video: tegra: edid: Fix computing of eld.sad_count
eld.sad_count as used in hda_eld.c is the count of short
audio descriptors in the edid, each of which contains three
bytes of data.  the old code was assigning the number of
bytes instead of the count.

Change-Id: I161b14fa8f61ff611614f41a3898ae3687fc1410
Signed-off-by: Mike J. Chen <mjchen@google.com>
Signed-off-by: Xia Yang <xiay@nvidia.com>
Signed-off-by: Jon Mayo <jmayo@nvidia.com>
Reviewed-on: http://git-master/r/354009
(cherry picked from commit 25f131743e6acfa980e1ce4b457c30891b6044bb)
Reviewed-on: http://git-master/r/383727
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Alex Courbot <acourbot@nvidia.com>
Tested-by: Alex Courbot <acourbot@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>