video: tegra: hdmi: fix audio enabling on boot
When hwc doesn't blank/unblank at boot because the mode it wants
is already set, audio wasn't working because at boot, the
code that sets the hdmi_audio switch is invoked by
tegra_dc_hdmi_enable(), but tegra_dc_hdmi_enable() runs before
EDID has been read.  That order is reversed in the hotplug case,
which is why it works there.

The fix is to split out the setup of audio and infoframes to a
function.  tegra_dc_hdmi_enable() only calls that function when
eld_retrieved is true, which should be the case for hotplug.
It doesn't call it when at boot.  Instead, we have the hdmi state
machine worker thread call it if the hdmi clk is enabled.  The only
time the clk should be enabled is in the boot case (because hdmi
enable was called first).  In hotplug case, hdmi was disabled, and
isn't enabled until after EDID has been read.

This change also adds a lock to eld_retrieved since it's set in
the worker thread and usually read in another thread.

Bug 1401958
Bug 1425201

Change-Id: Iee131f45a97d2466f03b9df0e1f4c967132e72b6
Signed-off-by: Mike J. Chen <mjchen@google.com>
Signed-off-by: Xia Yang <xiay@nvidia.com>
Reviewed-on: http://git-master/r/355342
(cherry picked from commit 478b1ee2950b0d64215f1dc849005064f883a964)
Reviewed-on: http://git-master/r/383737
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Alex Courbot <acourbot@nvidia.com>
Tested-by: Alex Courbot <acourbot@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>