media: rc: Don't depend input repeat for decoding
REP_DELAY is intended for the delay time before keyboard
driver start to fire repeating keyevents. On some systems
like Android, they are all set to be 0 as kernel repeating
is not needed. While this is fine for keyboard drivers, it
is devastating for ir-raw driver, as it use REP_DELAY as a
reference for IR pulse/space duration.

Bug 1353511

Change-Id: I617c83cd557c3378328ec79347568e0052422433
Signed-off-by: Jun Yan <juyan@nvidia.com>
Reviewed-on: http://git-master/r/391582
GVS: Gerrit_Virtual_Submit
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>
Reviewed-by: Raymond Poudrier <rapoudrier@nvidia.com>