video: tegra: dc: use override modes for probe only
Override modes (i.e. modes already set up by the bootloader and which
the kernel is supposed by keep using) should not be used in a hotplug
scenario, as nothing can tell us that the plugged display is the same as
the one the device booted with and thus should use the same modes.

Disable the use of override modes at the end of the DC probe function
to make sure this does not happen. Without this tegra_dc_program_mode()
does not get called during hotplug.

Bug 1425201

Change-Id: Iebf9dc3647fca2956458653375362706fdcee100
Signed-off-by: Alexandre Courbot <acourbot@nvidia.com>
Reviewed-on: http://git-master/r/387570
Tested-by: Vinayak Pane <vpane@nvidia.com>
Reviewed-by: Jon Mayo <jmayo@nvidia.com>
Reviewed-by: Tao Xie <txie@nvidia.com>
Reviewed-by: Vinayak Pane <vpane@nvidia.com>