video: tegra: dc: correctly use CMU cache
_tegra_dc_update_cmu() compares the CMU cache with the new settings by
comparing the addresses of both tegra_dc_cmu structures, but the cache
actually holds a *copy* of the CMU settings, so the comparision will
always be false. Fix this by using memcmp().

Also clear the cache when suspending to force CMU reprogramming upon
resume, as CMU settings are lost by suspend. Without this patch the
cache is effectively ignored so the need for cache reset remained
undetected.

Bug 1425201
Bug 1496993

Change-Id: I14f3505d1ccc3912f3a417eb1edf97654dc9beb1
Signed-off-by: Alexandre Courbot <acourbot@nvidia.com>
Reviewed-on: http://git-master/r/392841
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Jon Mayo <jmayo@nvidia.com>
Reviewed-by: Tao Xie <txie@nvidia.com>
Reviewed-by: Vinayak Pane <vpane@nvidia.com>
GVS: Gerrit_Virtual_Submit