media: tegra: nvavp: Fix corruption in secure playback after lp0
If vde partition is powergated before suspend, access to BSEV registers
that happens immediately after lp0 exit will go wrong. This WAR unpowergate
vde partition in suspend callback and powergate it again in the first place
of resume callback so that we won't keep vde powergate always on. Resume of
vde if needed will be taken care of afterwards.

Bug 1502013

Change-Id: Iaf4c36ac377c64e8d09d56f79e6e946ec47c686b
Signed-off-by: Allen Yu <alleny@nvidia.com>
Reviewed-on: http://git-master/r/401252
Tested-by: Praveen Kumar Reddy M.V. <pkreddy@nvidia.com>
Reviewed-by: Hunk Lin <hulin@nvidia.com>
Reviewed-by: Todd Poynter <tpoynter@nvidia.com>