net: wireless: bcmdhd: handle roam indication
When user modifies any settings in AP, wlan driver
does a reconnection on its own and after connection
is re-established, reports it as ROAM_EVENT to cfg.
Sometimes bss information of that particular bssid
gets expired in the cfg80211 which is resulting in
the WARN_ON in cfg80211_roamed function.
So skip roaming indication to cfg80211 layer if it
is connected to the same BSSID

Bug 1500472

Change-Id: I67b9466840451d65a3d7558fc14294ecf1ccf865
Signed-off-by: Narayan Reddy <narayanr@nvidia.com>
Reviewed-on: http://git-master/r/403994
Reviewed-by: Tony Ly <tly@nvidia.com>
Tested-by: Tony Ly <tly@nvidia.com>