tn8: Make backlight nonlinear
As per the request, we would want to make the backlight nonlinear in
progression. Do this as a step in the kernel before applying SD.
The curve implemented is a simple piecewise linear function with
a slow rise from 0 to 166 and a linear function with a faster rise
from 166 to 255.

Bug 200008701

Change-Id: I9dfdf17c448165a1b5a770a0db7f4c9beded4439
Signed-off-by: Aly Hirani <ahirani@nvidia.com>
Reviewed-on: http://git-master/r/418526
Reviewed-by: Simone Willett <swillett@nvidia.com>
Tested-by: Simone Willett <swillett@nvidia.com>