video: tegra: host: Do not deinit tsec
tsec deinitialisation is currently triggered when the last user
on tsec channel closes the channel and we release the firmware as
part of the deinitialisation routine. However, we may still have some
work pending on the channel and therefore releasing the firmware
may trigger unexpected faults from tsec.

As firmware release is not needed (i.e. we will need the firmware)
later when the channel is used next time, we can simply remove the
firmware release.

Bug 200018717

Change-Id: I0b2c381049c6dbec43be8f42a670fdd7b06ff400
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/435056
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Dhiren Parmar <dparmar@nvidia.com>
Tested-by: Dhiren Parmar <dparmar@nvidia.com>