staging: iio: light: iqs253: make sensor always on
Keep sensor always on for its proper functioning.
It seems, we do not work fine with sensor turning on when
needed and turning it off when not in use. This behavior
could be because of the time it takes to adjust to the
environment but the root cause is not yet clear.

Bug 1499925

Change-Id: I57a5949e01f8d9cf68e50d877a2dae6925592ffd
Signed-off-by: Sri Krishna chowdary <schowdary@nvidia.com>
Reviewed-on: http://git-master/r/436506
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Chas Inman <chasi@nvidia.com>