net:wireless:bcmdhd: Don't enable Mon mode for p2p
P2P CERT 6.1.9 test case is failing becasue monitor mode
is creating an issue in registering Action frame for P2P-GO in
firmware, this was leading an error in receiving action frames
to GO interface. Not enabling it for P2P in driver though
DUT does not support the mode.

Bug 200049414

Change-Id: I296f0c9d799d6978e946399ccc1f0da1b6d31d8d
Signed-off-by: Raghavender <kraghavender@nvidia.com>
Reviewed-on: http://git-master/r/590780
Reviewed-by: Mohan Thadikamalla <mohant@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Todd Poynter <tpoynter@nvidia.com>