PM/Suspend:suspend_entry_time population corrected
suspend_entry_time variable was only getting populated when
device's display goes off and not when modules are getting
suspended for while device tries to go in LP0.
Also, the point at which suspend_entry_time variable
gets populated is changed to the first function which gets
triggered while the device tries to go in LP0.

Bug 200044223

Change-Id: I7f3d3eb4cfae16855030d92601d209240ec5e916
Signed-off-by: Saket Bhojane <sbhojane@nvidia.com>
Reviewed-on: http://git-master/r/606860
Reviewed-by: Dhiren Parmar <dparmar@nvidia.com>
Tested-by: Dhiren Parmar <dparmar@nvidia.com>