arm: tegra: tn8: enable CMU for HDMI
As per the HDMI standard, we should be defaulting to the defined
"limited" color range. In Tegra, we use CMU on dc1/HDMI to achieve
this.

In the userspace, NvCplService is responsible for setting the cmu_enable
node as per the user configuration. However, it actually picks up the
default setting based on what the kernel sets it.

Prior to this change, we were actually defaulting to cmu_enable being
off. This would actually cause us to send full range data with the
limited range avi infoframe and hence the "output is too dark" concern.
This change enables the cmu by default on HDMI instead.

Bug 1572199

Change-Id: I738b2aea007035c688ed091a212a45074e617c53
Signed-off-by: Aly Hirani <ahirani@nvidia.com>
Reviewed-on: http://git-master/r/660563
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Min-wuk Lee <mlee@nvidia.com>
Reviewed-by: Todd Poynter <tpoynter@nvidia.com>