video: tegra: dc: remove "dc->enable" check in "tegra_dc_update_cmu_aligned"
On bootup, NvCPL will phase-in user-set CMU change.
If the user disables the panel during that first
phase-in, ioctls to update CMU after DC gets
disabled will be lost.

To avoid this we simply do not check whether DC
is enabled, but go ahead and cache the data from
the ioctl anyway. Next time DC is enabled, CMU
will be updated with the most recently cached data.

Change-Id: I6144ae306dffbc895d00e861b0c643cc564f8594
Signed-off-by: Arun Swain <arswain@nvidia.com>
Reviewed-on: http://git-master/r/664646
Reviewed-by: Daniel Solomon <daniels@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Mitch Luban <mluban@nvidia.com>