net: wireless: bcmdhd: Add support for region based country code
Earlier the Q2/87 country code was used in the nvram to support a
global list of channels in the wifi firmware. The txpower was also
controlled in the firmware.

To allow better usage of DFS channels depending on the country and
allow higher txpower depending on channel and operating band, nvram
is finely tuned and hardcoded in the driver for the following country
codes - JP/89, KR/89, RU/59, TW/62, XV/32, E0/44, US/0. Depending on
the actual geo-location of the STA, the country code is dynamically
changed and firmware is reloaded for changes to take effect.

Bug 1553567

Change-Id: I702527340ddf96c4aeb28a07d3ce969c944aab6e
Signed-off-by: Srinivas Ramachandran <srinivasra@nvidia.com>
Signed-off-by: Om Prakash Singh <omp@nvidia.com>
Reviewed-on: http://git-master/r/669103
Reviewed-by: Dhiren Parmar <dparmar@nvidia.com>