video: tegra: dc: fix race condition in cmu setting
When CMU is first set by DC after resume, we use
_tegra_dc_update_cmu_aligned with force = true, to
force writing of all registers.
Since that write occurs on the next end of frame, it's
possible for a cmu_aligned ioctl to come in beforehand
and overwrite force = false, which will prevent setting
of some registers.

This change takes into consideration the current
status of the force flag when setting that flag due
to new requests.

Bug 1598319

Change-Id: I9d6f55db6faaad5db5107400d224e3ca66366c73
Signed-off-by: Daniel Solomon <daniels@nvidia.com>
Reviewed-on: http://git-master/r/672521
Reviewed-by: Aly Hirani <ahirani@nvidia.com>
Reviewed-by: Mitch Luban <mluban@nvidia.com>