misc: tegra-baseband: remove remote wakeup callback
The ehci-tegra driver is holding a wakelock as long as the bus is active
so the wakelock taken upon remote wake in tegra_usb_modem_power is no
longer necessary.

Additionally, ehci-tegra also skips resume of the modem if it was suspended
prior to LP0 entry so the short autosuspend delay for power savings is no
no longer necessary.

Bug 1557355

Change-Id: I67fe998ef5cff7fac9282d20338b31ee50a42ba1
Signed-off-by: Neil Patel <neilp@nvidia.com>
Reviewed-on: http://git-master/r/681638
Reviewed-by: Rajkumar Jayaraman <rjayaraman@nvidia.com>
Reviewed-by: Steve Lin <stlin@nvidia.com>
Reviewed-by: Mark Kuo <mkuo@nvidia.com>
Reviewed-by: Dhiren Parmar <dparmar@nvidia.com>